Move some methods from DynamicSimulationService to DynamicSimulationResultService
